requirements.txt
langchain==0.3.27
langchain_community==0.3.27
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.48.1
